Move the BILL_STATUS and HOLIDAY_TYPES contants to constans/index file
