--- app/javascript/src/components/TimeoffEntries/constants.ts
@@ -1,10 +1,0 @@
-export const BILL_STATUS = {
-  BILLED: "billed",
-  UNBILLED: "unbilled",
-  NON_BILLABLE: "non_billable",
-};
-
-export const HOLIDAY_TYPES = {
-  NATIONAL: "national",
-  OPTIONAL: "optional",
-};
